Update README.md with notes on creating subdirectories for packs
I have been splitting up my packs by system, so I thought it would be helpful to share with others how to do that. 

It is entirely possible that this is not a good idea. If so, feel free to reject and let me know how it should be done instead, and I can update this.