Add --config command line flag
When run with --config=/path/to/config.yml, read the configuration
values from that file and merge them into the values from the global
config file. If a configuration value is set using the 'configure set'
command, the value will be written to the local config file specified
instead of the global file.

Signed-off-by: Joe Nosie <182554+BoltsJ@users.noreply.github.com>