Add ability to pack objects rather than just files
This allows compilePack() src argument to not just be a directory
with files to pack, but an Iterable that returns the objects to pack.

Then it can be used to create packs that are dynamically created.

PF2e Workbench can use this for a compendium of macros it creates, where the
macros don't exist as directory of json files.  The macro Items are created
dynamically.  Currently, it uses an old NeDB format compendium for these and
makes the database "manually" by concatenating lines of json text.