[#5] Add option to extract with matching folder structure
Adds the `folders` option to extraction that places extracted
documents into folders matching the folders in the original
compendium. The folder data itself is added to `_Folder.json/yml`
within the created folder.

Adds a `transformFolderName` option when programmatically calling
`extractPack` that allows for renaming the folder that will be
created. The normal `transformName` callback also gets a second
parameter with the containing folder name.

Closes foundryvtt/foundryvtt-cli#5