[#46] Extract documents within adventures into separate files
Adds a new `expandAdventures` option that extracts all of the
documents within an adventure into their own files when extracting
from LevelDB. When this option is selected each individual
document in the adventure is extracted as if they were primary
documents and gets `transformEntry` and `transformName` called on
them with an additional context object that contains information
on the adventure itself. The adventure is then saved with the
filenames for the written files in place of the whole objects.

When compiling back to NeDB or LevelDB, the CLI will check for any
strings within the adventure document and reassemble it from the
files listed.

Closes #46