[#5] Add option to extract with matching folder structure (#60)
* [#5] Add option to extract with matching folder structure

Adds the `folders` option to extraction that places extracted
documents into folders matching the folders in the original
compendium. The folder data itself is added to `_Folder.json/yml`
within the created folder.

Adds a `transformFolderName` option when programmatically calling
`extractPack` that allows for renaming the folder that will be
created. The normal `transformName` callback also gets a second
parameter with the containing folder name.

Closes foundryvtt/foundryvtt-cli#5

* [#5] Move folder passed to `NameTransformer` into its own object

Moved the `folder` passed into `NameTransformer` into a `context`
object to potentially support additional context in the future.

* [#5] Fix bug if folders setting isn't enabled

* [#5] Pass `folders` argument in from command line properly

* Add docs. Update CHANGELOG.

---------

Co-authored-by: fyorl <kim.mantas@gmail.com>