Fix package not being read from local config
Because the currentPackageId and currentPackageType variables were being
set in the top level of the package module, they were being set before
any specified local config was loaded.  This moves the read from Config
from the declaration to the top of command handler, ensuring the final
value from the combined config is loaded.